refactor the project's directory structure
