+++ perf/perf-producer.go
@@ -22,7 +22,6 @@ package main
 import (
 	"context"
 	"encoding/json"
-	`github.com/Anankke/bilibili-backup/library/net/http/blademaster/middleware/rate`
 	"github.com/beefsack/go-rate"
 	"github.com/bmizerany/perks/quantile"
 	"github.com/spf13/cobra"
