+++ pkg/compression/compression.go
