+++ pkg/compression/compression_test.go
