+++ pkg/compression/lz4.go
