+++ pulsar/impl_message.go
@@ -21,7 +21,7 @@ package pulsar
 
 import (
 	"github.com/golang/protobuf/proto"
-	pb "github.com/apache/pulsar-client-go/pulsar/internal/pulsar_proto"
+	"github.com/apache/pulsar-client-go/pkg/pb"
 )
 
 type messageId struct {
