+++ pulsar/impl_partition_producer.go
@@ -24,8 +24,8 @@ import (
 	"github.com/golang/protobuf/proto"
 	log "github.com/sirupsen/logrus"
 	"github.com/apache/pulsar-client-go/pulsar/internal"
-	"github.com/apache/pulsar-client-go/pulsar/internal/util"
-	pb "github.com/apache/pulsar-client-go/pulsar/internal/pulsar_proto"
+	"github.com/apache/pulsar-client-go/util"
+	"github.com/apache/pulsar-client-go/pkg/pb"
 	"sync"
 	"sync/atomic"
 	"time"
