+++ pulsar/internal/connection.go
@@ -24,12 +24,12 @@ import (
 	"crypto/x509"
 	"errors"
 	"github.com/golang/protobuf/proto"
-	log "github.com/sirupsen/logrus"
+    log "github.com/sirupsen/logrus"
 	"io/ioutil"
 	"net"
 	"net/url"
-	"github.com/apache/pulsar-client-go/pulsar/internal/auth"
-	pb "github.com/apache/pulsar-client-go/pulsar/internal/pulsar_proto"
+	"github.com/apache/pulsar-client-go/pkg/auth"
+	"github.com/apache/pulsar-client-go/pkg/pb"
 	"sync"
 	"sync/atomic"
 	"time"
